chore: remove v2 references from operation data access service and its usages
